+++ core/api/src/main/java/app/kaito_dogi/mybrary/core/api/google/GoogleApiModule.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.api.googlebooks
+package app.kaito_dogi.mybrary.core.api.google
 
 import dagger.Module
 import dagger.Provides
@@ -11,21 +11,21 @@ import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.kotlinx.serialization.asConverterFactory
 
-private const val GoogleBooksBaseUrl = "https://www.googleapis.com/"
+private const val GoogleApiBaseUrl = "https://www.googleapis.com/"
 
 @Module
 @InstallIn(SingletonComponent::class)
-internal object GoogleBooksApiModule {
+internal object GoogleApiModule {
   @Singleton
   @Provides
-  fun provideGoogleBooksApi(okHttpClient: OkHttpClient): GoogleBooksApi {
+  fun provideGoogleApi(okHttpClient: OkHttpClient): GoogleApi {
     val contentType = "application/json; charset=UTF8".toMediaType()
     val format = Json { ignoreUnknownKeys = true }
     return Retrofit.Builder()
       .client(okHttpClient)
-      .baseUrl(GoogleBooksBaseUrl)
+      .baseUrl(GoogleApiBaseUrl)
       .addConverterFactory(format.asConverterFactory(contentType))
       .build()
-      .create(GoogleBooksApi::class.java)
+      .create(GoogleApi::class.java)
   }
 }
