+++ core/api/src/main/java/app/kaito_dogi/mybrary/core/api/google/response/GetVolumesResponse.kt
@@ -1,0 +1,9 @@
+package app.kaito_dogi.mybrary.core.api.google.response
+
+import app.kaito_dogi.mybrary.core.api.google.response.model.ItemResponse
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GetVolumesResponse(
+  val items: List<ItemResponse>,
+)
