+++ core/api/src/main/java/app/kaito_dogi/mybrary/core/api/google/response/model/ItemResponse.kt
@@ -1,4 +1,4 @@
-package app.kaito_dogi.mybrary.core.api.googlebooks.response.model
+package app.kaito_dogi.mybrary.core.api.google.response.model
 
 import kotlinx.serialization.Serializable
